requirements.txt
# ============================================================
# Cross-Selling Recommender System — Dependencies
# Proyecto Final | Data Science — Henry
# ============================================================

# --- Core Data ---
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0             # Matrices sparse CSR para ALS

# --- EDA & Visualization ---
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# --- Machine Learning ---
scikit-learn==1.4.2

# --- Recommendation Models ---
mlxtend==0.23.1           # Apriori / FP-Growth / Association Rules
implicit==0.7.2           # ALS (implicit feedback — feedback implícito)
scikit-surprise==1.1.4    # SVD (opcional, para comparación con ALS)

# --- Experiment Tracking ---
mlflow==2.13.0

# --- API & Deployment ---
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0

# --- Utilities ---
python-dotenv==1.0.1
pyyaml==6.0.1
joblib==1.4.2
tqdm==4.66.4

# --- Testing ---
pytest==8.2.0
pytest-cov==5.0.0
